+++ src/algorithms/flow/mod.rs
@@ -11,5 +11,7 @@
 //! ones.
 
 pub(crate) mod max_flow;
+pub(crate) mod st_mincut;
 
 pub use max_flow::max_flow_value;
+pub use st_mincut::st_mincut_value;